+++ opercam/src/main/java/com/app/opercam/model/repository/IRechazoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.app.opercam.model.entity.Rechazo;
 
-public interface IRechazoRepository extends CrudRepository<Rechazo, Integer> {
+public interface IRechazoRepository extends CrudRepository<Rechazo, Long> {
 
 }
